Add match diagnostic primary ranges
