+++ crates/sifr/tests/e2e/fail/match_invalid_field_name.sifr
@@ -1,4 +1,4 @@
-# expect-error: SIFR-MATCH-0003
+# expect-error[col=26]: SIFR-MATCH-0003
 class Point:
     x: int
     y: int
