+++ crates/sifr/tests/e2e/fail/match_non_exhaustive_union.sifr
@@ -1,4 +1,4 @@
-# expect-error: SIFR-MATCH-0001
+# expect-error[col=11]: SIFR-MATCH-0001
 def describe(x: int | None) -> str:
     match x:
         case int():
