+++ issues/ad-hoc-semantic-diagnostic-code-taxonomy-and-structured-hir-diagnostics.md
@@ -107,6 +107,8 @@ Current wave: `milestone_diag_9` source span completion, split into reviewable s
 - [x] Claude implementation review for `milestone_diag_9` slice 6 completed and reviewer is satisfied. Review rounds: `reviews/semantic-diagnostic-code-taxonomy-diag-9-ownership-primary-ranges-review-pass-1.md`, `reviews/semantic-diagnostic-code-taxonomy-diag-9-ownership-primary-ranges-review-pass-2.md`. Local validation passed: `cargo fmt --check`, `python3 scripts/check_hir_maintainability_guardrails.py`, `cargo test -p sifr_hir own_mut_semantics_tests -- --nocapture`, focused HIR ownership expression primary-range tests, `cargo test -p sifr --test e2e test_e2e_fail -- --nocapture`, `cargo clippy -p sifr_hir --no-deps -- -D warnings`, and `scripts/run_all_tests.sh --profile quick` (`report_signature=e1bf653aaa770517`, `wall_time=52.94s`; group-skew advisory emitted).
 - [x] `milestone_diag_9` slice 7 implementation complete and reviewer-satisfied: attached primary ranges to remaining flow diagnostics (`SIFR-FLOW-0003`, `SIFR-FLOW-0004`, and the residual `SIFR-FLOW-0005` condition helper), removed the optional span fallback from condition validation, and added HIR/e2e column coverage for nonlocal and missing-return flow paths. PR: https://github.com/sifr-lang/sifr/pull/1731.
 - [x] Claude implementation review for `milestone_diag_9` slice 7 completed and reviewer is satisfied. Review round: `reviews/semantic-diagnostic-code-taxonomy-diag-9-flow-remaining-primary-ranges-review-pass-1.md`. Local validation passed: `cargo fmt --check`, `python3 scripts/check_hir_maintainability_guardrails.py`, `cargo test -p sifr_hir condition_rejects_numeric_truthiness -- --nocapture`, `cargo test -p sifr_hir nested_function_tests -- --nocapture`, `cargo test -p sifr_hir test_non_none_return_annotation_requires_exhaustive_returns -- --nocapture --test-threads=1`, `cargo test -p sifr --test e2e test_e2e_fail -- missing_return_value nested_function_recursive_nonlocal_unsupported elif_condition_numeric_truthiness --nocapture`, `cargo clippy -p sifr_hir --no-deps -- -D warnings`, and `scripts/run_all_tests.sh --profile quick` (`report_signature=e1bf653aaa770517`, `wall_time=52.13s`; group-skew advisory emitted).
+- [x] `milestone_diag_9` slice 8 implementation complete and reviewer-satisfied: attached primary ranges to match diagnostics (`SIFR-MATCH-0001` through `SIFR-MATCH-0003`), split match-statement lowering out of `statements.rs` to satisfy HIR maintainability guardrails, and added HIR primary-range assertions plus e2e column anchors for match fixtures. PR: pending.
+- [x] Claude implementation review for `milestone_diag_9` slice 8 completed and reviewer is satisfied. Review round: `reviews/semantic-diagnostic-code-taxonomy-diag-9-match-primary-ranges-review-pass-1.md`. Local validation passed: `cargo fmt --check`, `git diff --check`, `python3 scripts/check_hir_maintainability_guardrails.py`, `cargo test -p sifr_hir match_diagnostics -- --nocapture --test-threads=1`, `cargo test -p sifr --test e2e test_e2e_fail -- match_non_exhaustive_optional match_invalid_field_name enum_match_non_exhaustive match_non_exhaustive_union match_type_mismatch_guard match_non_exhaustive_literal --nocapture`, `cargo clippy -p sifr_hir --no-deps -- -D warnings`, and `scripts/run_all_tests.sh --profile quick` (`report_signature=e1bf653aaa770517`, `wall_time=62.80s`; warm-cache hit-rate and group-skew advisories emitted).
 - [x] Claude pre-implementation review for `milestone_diag_4a` slice 2 completed: `reviews/semantic-diagnostic-code-taxonomy-diag-4a-slice2-preimplementation-review-pass-1.md`.
 - [x] Claude implementation review for `milestone_diag_4a` slice 2a completed and all actionable findings addressed. Review rounds: `reviews/semantic-diagnostic-code-taxonomy-diag-4a-slice2a-transport-review-pass-1.md`, `reviews/semantic-diagnostic-code-taxonomy-diag-4a-slice2a-transport-review-pass-2.md`.
 - [x] Claude implementation review for `milestone_diag_4a` slice 2b.1 completed and reviewer is satisfied. Review rounds: `reviews/semantic-diagnostic-code-taxonomy-diag-4a-decimal-review-pass-1.md`, `reviews/semantic-diagnostic-code-taxonomy-diag-4a-decimal-review-pass-2.md`.
